+++ ImageFeed/Service Access Layer/Profile Service/ProfileResultModel.swift
@@ -19,3 +19,5 @@ struct ProfileImage: Codable {
         let medium: String
         let large: String
 }
+
+
