+++ ImageFeed/Service Access Layer/Profile Service/ProfileViewModel.swift
@@ -6,8 +6,20 @@
 //
 
 import Foundation
+
 struct Profile {
-    let username = {
-        
+    let username: String
+    let firstName: String
+    let lastName: String
+    
+    var name: String {
+        return "\(firstName) \(lastName)"
     }
+    
+    var loginName: String {
+        return "@\(username)"
+    }
+    
+    let bio: String
 }
+
